+++ ReleaseNotes.md
@@ -1,5 +1,6 @@
 ### New in 2.0.6
 * Added support to store original or set new resolution (DPI) 
+* Added RotateFlip method
 
 ### New in 2.0.5
 * Rewritten file loader to prevent file locks